add scUSDCv2 strategy (#58)
Add scUSDCv2: Deposit USDC and have it over-collateralize a deposit into scWETHv2 by taking on WETH debt on multiple lending protocols. Receive scUSDCv2 shares which appreciate over time when the vault is profitable and can be redeemed for USDC. A keeper bot regularly rebalances the positions against a target loan to value ratio and allocation configuration in order to compound profits or de-risk positions. Strategy is implemented as a composable ERC4626 that can interact with an arbitrary number of lending protocols thru the use of adapters. Initial version has exposure to multiple protocols such as scWETHv2, Balancer, Aave V2/V3, Uniswap V3 and the Chainlink USDC/ETH price feed.

---------

Co-authored-by: coolhill <8256597+coolhill@users.noreply.github.com>